requirements.txt
Django==1.7.1
djangorestframework==3.0.0
inflection==0.2.1
